LayoutTests/platform/win/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange-expected.txt



PASS test of input.setSelectionRange
PASS input typeof(input.setSelectionRange)'
PASS input setSelectionRange return void
PASS input setSelectionRange(0,1)
PASS input setSelectionRange(0,input.value.length+1)
PASS input setSelectionRange(2,2)
PASS input setSelectionRange(2,1)
PASS input direction of setSelectionRange(0,1,"backward")
PASS input direction of setSelectionRange(0,1,"forward")
FAIL input direction of setSelectionRange(0,1,"none") assert_equals: The direction of the selection must be set to forward if direction is a case-sensitive match for the string "none" expected "none" but got "forward"
FAIL input direction of setSelectionRange(0,1,"hoge") assert_equals: otherwise expected "none" but got "forward"
FAIL input direction of setSelectionRange(0,1,"BACKWARD") assert_equals: selectionDirection should be 'none' expected "none" but got "forward"
FAIL input direction of setSelectionRange(0,1) assert_equals: if the argument is omitted expected "none" but got "forward"
PASS input setSelectionRange("string",1)
PASS input setSelectionRange(true,1)
PASS input setSelectionRange([],1)
PASS input setSelectionRange({},1)
PASS input setSelectionRange(NaN,1)
PASS input setSelectionRange(null,1)
PASS input setSelectionRange(undefined,1)
PASS test of textarea.setSelectionRange
PASS textarea typeof(input.setSelectionRange)'
PASS textarea setSelectionRange return void
PASS textarea setSelectionRange(0,1)
PASS textarea setSelectionRange(0,textarea.value.length+1)
PASS textarea setSelectionRange(2,2)
PASS textarea setSelectionRange(2,1)
PASS textarea direction of setSelectionRange(0,1,"backward")
PASS textarea direction of setSelectionRange(0,1,"forward")
FAIL textarea direction of setSelectionRange(0,1,"none") assert_equals: The direction of the selection must be set to forward if direction is a case-sensitive match for the string "none" expected "none" but got "forward"
FAIL textarea direction of setSelectionRange(0,1,"hoge") assert_equals: otherwise expected "none" but got "forward"
FAIL textarea direction of setSelectionRange(0,1,"BACKWARD") assert_equals: selectionDirection should be 'none' expected "none" but got "forward"
FAIL textarea direction of setSelectionRange(0,1) assert_equals: if the argument is omitted expected "none" but got "forward"
FAIL textarea setSelectionRange("string",1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
PASS textarea setSelectionRange(true,1)
FAIL textarea setSelectionRange([],1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
FAIL textarea setSelectionRange({},1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
FAIL textarea setSelectionRange(NaN,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
FAIL textarea setSelectionRange(null,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
FAIL textarea setSelectionRange(undefined,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0
